Refactor draw.js into one module per draw function
