+++ js-lib/draw/index.js
@@ -1,0 +1,9 @@
+export { default as DrawFunction } from './function';
+export { default as DrawBox } from './box';
+export { default as DrawBoxText } from './boxtext';
+export { default as DrawErase } from './erase';
+export { default as DrawLine } from './line';
+export { default as DrawSelect } from './select';
+export { default as DrawText } from './text';
+export { default as DrawMove } from './move';
+export { default as DrawFreeform } from './freeform';
